int to short to kinda reduce memory

+++ Obsidian.API/_Interfaces/IServer.cs
@@ -10,7 +10,7 @@ namespace Obsidian.API
     {
         public string Version { get; }
         public int Port { get; }
-        public short TPS { get; }
+        public byte TPS { get; }
         public DateTimeOffset StartTime { get; }
         public ProtocolVersion Protocol { get; }
         public IEnumerable<IPlayer> Players { get; }
