+++ Obsidian/Blocks/BlockDirtSnow.cs
@@ -2,6 +2,6 @@
 {
     public class BlockDirtSnow : BlockGrass
     {
-        internal BlockDirtSnow(string name, int id, Materials type) : base(name, id, type) { }
+        internal BlockDirtSnow(string name, short id, Materials type) : base(name, id, type) { }
     }
 }
