+++ Obsidian/Blocks/BlockDispenser.cs
@@ -2,7 +2,7 @@
 {
     public class BlockDispenser : Block
     {
-        internal BlockDispenser(string name, int id, Materials type) : base(name, id, type)
+        internal BlockDispenser(string name, short id, Materials type) : base(name, id, type)
         {
         }
     }
