+++ Obsidian/Blocks/BlockEnchantingTable.cs
@@ -2,7 +2,7 @@
 {
     public class BlockEnchantingTable : Block
     {
-        internal BlockEnchantingTable(string name, int id) : base(name, id, Materials.EnchantingTable)
+        internal BlockEnchantingTable(string name, short id) : base(name, id, Materials.EnchantingTable)
         {
         }
     }
