+++ Obsidian/Blocks/BlockFurnace.cs
@@ -2,7 +2,7 @@
 {
     public class BlockFurnace : Block
     {
-        internal BlockFurnace(string name, int id) : base(name, id, Materials.Furnace)
+        internal BlockFurnace(string name, short id) : base(name, id, Materials.Furnace)
         {
         }
     }
