+++ Obsidian/Blocks/BlockHugeMushroom.cs
@@ -2,7 +2,7 @@
 {
     public class BlockHugeMushroom : Block
     {
-        internal BlockHugeMushroom(string name, int id, Materials type) : base(name, id, type)
+        internal BlockHugeMushroom(string name, short id, Materials type) : base(name, id, type)
         {
         }
     }
