+++ Obsidian/Blocks/BlockIronBars.cs
@@ -2,7 +2,7 @@
 {
     public class BlockIronBars : Block
     {
-        internal BlockIronBars(string name, int id) : base(name, id, Materials.IronBars)
+        internal BlockIronBars(string name, short id) : base(name, id, Materials.IronBars)
         {
         }
     }
