+++ Obsidian/Blocks/BlockMelon.cs
@@ -2,7 +2,7 @@
 {
     public class BlockMelon : Block
     {
-        internal BlockMelon(string name, int id) : base(name, id, Materials.Melon)
+        internal BlockMelon(string name, short id) : base(name, id, Materials.Melon)
         {
         }
     }
