+++ Obsidian/Blocks/BlockMonsterEggs.cs
@@ -2,7 +2,7 @@
 {
     public class BlockMonsterEggs : Block
     {
-        internal BlockMonsterEggs(string name, int id, Materials type) : base(name, id, type)
+        internal BlockMonsterEggs(string name, short id, Materials type) : base(name, id, type)
         {
         }
     }
