+++ Obsidian/Blocks/BlockOre.cs
@@ -2,7 +2,7 @@
 {
     public class BlockOre : Block
     {
-        internal BlockOre(string name, int id, Materials type) : base(name, id, type)
+        internal BlockOre(string name, short id, Materials type) : base(name, id, type)
         {
         }
     }
