+++ Obsidian/Blocks/BlockPistonExtension.cs
@@ -2,7 +2,7 @@
 {
     public class BlockPistonExtension : BlockPiston
     {
-        internal BlockPistonExtension(string name, int id, Materials type) : base(name, id, type)
+        internal BlockPistonExtension(string name, short id, Materials type) : base(name, id, type)
         {
         }
     }
