+++ Obsidian/Blocks/BlockPressurePlate.cs
@@ -2,7 +2,7 @@
 {
     public class BlockPressurePlate : Block
     {
-        internal BlockPressurePlate(string name, int id, Materials type) : base(name, id, type)
+        internal BlockPressurePlate(string name, short id, Materials type) : base(name, id, type)
         {
         }
     }
