+++ Obsidian/Blocks/BlockPumpkinCarved.cs
@@ -2,7 +2,7 @@
 {
     public class BlockPumpkinCarved : Block
     {
-        internal BlockPumpkinCarved(string name, int id, Materials type) : base(name, id, type)
+        internal BlockPumpkinCarved(string name, short id, Materials type) : base(name, id, type)
         {
         }
     }
