+++ Obsidian/Blocks/BlockRedstoneWire.cs
@@ -2,7 +2,7 @@
 {
     public class BlockRedstoneWire : Block
     {
-        internal BlockRedstoneWire(string name, int id) : base(name, id, Materials.RedstoneWire)
+        internal BlockRedstoneWire(string name, short id) : base(name, id, Materials.RedstoneWire)
         {
         }
     }
