+++ Obsidian/Blocks/BlockSeaGrass.cs
@@ -2,7 +2,7 @@
 {
     public class BlockSeaGrass : Block
     {
-        internal BlockSeaGrass(string name, int id, Materials type) : base(name, id, type)
+        internal BlockSeaGrass(string name, short id, Materials type) : base(name, id, type)
         {
         }
     }
