+++ Obsidian/Blocks/BlockStairs.cs
@@ -2,7 +2,7 @@
 {
     public class BlockStairs : Block
     {
-        internal BlockStairs(string name, int id, Materials type) : base(name, id, type)
+        internal BlockStairs(string name, short id, Materials type) : base(name, id, type)
         {
         }
     }
