+++ Obsidian/Blocks/BlockStem.cs
@@ -2,7 +2,7 @@
 {
     public class BlockStem : Block
     {
-        internal BlockStem(string name, int id, Materials type) : base(name, id, type)
+        internal BlockStem(string name, short id, Materials type) : base(name, id, type)
         {
         }
     }
