+++ Obsidian/Blocks/BlockStemAttached.cs
@@ -2,7 +2,7 @@
 {
     public class BlockStemAttached : Block
     {
-        internal BlockStemAttached(string name, int id, Materials type) : base(name, id, type)
+        internal BlockStemAttached(string name, short id, Materials type) : base(name, id, type)
         {
         }
     }
