+++ Obsidian/Blocks/BlockWallSign.cs
@@ -2,7 +2,7 @@
 {
     public class BlockWallSign : Block
     {
-        internal BlockWallSign(string name, int id, Materials type) : base(name, id, type)
+        internal BlockWallSign(string name, short id, Materials type) : base(name, id, type)
         {
         }
     }
